feat: validation for input models
